move cli args into launcher module
